pyproject.toml
[tool.poetry]
name = "typerun"
version = "0.1.0"
description = ""
authors = ["Yryskeb <yryskeb9801@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.1"
psycopg2-binary = "^2.9.9"
python-decouple = "^3.8"
celery = "^5.3.6"
redis = "^5.0.1"
django-filter = "^23.5"
drf-yasg = "^1.21.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
